Changed MBytes and MBytes/sec to float.

+++ test.c
@@ -260,14 +260,14 @@ void run_tcp_client_test(void) {
 static void iperf_report(void *arg, enum lwiperf_report_type report_type,
                          const ip_addr_t *local_addr, u16_t local_port, const ip_addr_t *remote_addr, u16_t remote_port,
                          u32_t bytes_transferred, u32_t ms_duration, u32_t bandwidth_kbitpsec) {
-    static uint32_t total_iperf_megabytes = 0;
-    uint32_t mbytes = bytes_transferred / 1024 / 1024;
+    static float total_iperf_megabytes = 0.0;
+    float mbytes = bytes_transferred / 1024.0 / 1024.0;
     float mbits = bandwidth_kbitpsec / 1000.0;
 
     total_iperf_megabytes += mbytes;
 
-    printf("Completed iperf transfer of %d MBytes @ %.1f Mbits/sec\n", mbytes, mbits);
-    printf("Total iperf megabytes since start %d Mbytes\n", total_iperf_megabytes);
+    printf("Completed iperf transfer of %.2f MBytes @ %.2f Mbits/sec\n", mbytes, mbits);
+    printf("Total iperf megabytes since start %.2f Mbytes\n", total_iperf_megabytes);
 #if CYW43_USE_STATS
     printf("packets in %u packets out %u\n", CYW43_STAT_GET(PACKET_IN_COUNT), CYW43_STAT_GET(PACKET_OUT_COUNT));
 #endif
